--- tests/python/tensorrt/common.py
@@ -1,30 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-
-from ctypes.util import find_library
-
-
-def check_tensorrt_installation():
-    assert find_library('nvinfer') is not None, "Can't find the TensorRT shared library"
-
-
-def merge_dicts(*dict_args):
-    """Merge arg_params and aux_params to populate shared_buffer"""
-    result = {}
-    for dictionary in dict_args:
-        result.update(dictionary)
-    return result
